refactor(sumcheck): split svo.rs into a svo/ module by concern
The 2465-line `svo.rs` mixed four independent concerns plus the central
`SvoPoint` type. Split it into a `svo/` module (mod.rs style, matching
`layout/`) with one file per concern, no behavior change:

- `grid.rs`        ternary {0,1,inf}^l grid expansion (shared kernel)
- `accumulator.rs` product-accumulator kernels + per-round batch assembly
- `point.rs`       the `SvoPoint` type, whole (all inherent methods)
- `mod.rs`         module docs + re-exports (all `crate::svo::*` paths preserved)

The `l<=3` accumulator specializations now take fixed-size array refs
(`&[EF; N]`) instead of slices, moving the length check to the type system
and dropping the runtime asserts and index soup.

Relocate two functions to their owning types (they were misplaced):

- `next_state_evals` -> `NextPartials::from_point` in `layout/opening.rs`.
  It builds a `NextPartials` and was already called from inside that type's
  own `accumulate_*` methods.
- `next_round_partials_{suffix,prefix}` -> private `SvoPoint` methods (they
  were only ever called with `&self.z_svo`; `&Poly` args keep them clone-free).

This dissolves the weak `next.rs` module; its pieces live with their types.

Tests move with their code and are dispatched per file. No duplicate tests,
no dangling references. `cargo test -p p3-sumcheck` 220 passed; clippy clean.
